move adaptive methods to new wrapper
